Refactor to maintain single responsibility
